requirements.txt
pillow
scipy==1.0.1
matplotlib
imageio
sklearn
matplotlib
